🔨 Refactor utils
