+++ pages/api/login.ts
@@ -2,7 +2,7 @@
 
 import { withIronSessionApiRoute } from "iron-session/next";
 import { NextApiRequest, NextApiResponse } from "next";
-import { sessionOptions } from "../../utils/withSession";
+import { sessionOptions } from "utils";
 
 export default withIronSessionApiRoute(loginRoute, sessionOptions);
